+++ test/test_mysql.js
@@ -17,16 +17,14 @@ describe(__filename, function () {
     await assert.rejects(db.init());
   });
 
-  it('reconnect after fatal error', async function () {
+  it('query after fatal error works', async function () {
     const db = new mysql.Database(databases.mysql);
     await db.init();
-    const before = await db._connection;
     // An error is expected; prevent it from being logged.
     db.logger = Object.setPrototypeOf({error() {}}, db.logger);
     // Sleep longer than the timeout to force a fatal error.
     await assert.rejects(db._query({sql: 'DO SLEEP(1);', timeout: 1}), {fatal: true});
-    const after = await db._connection;
-    assert.notEqual(after, before);
+    await assert.doesNotReject(db._query({sql: 'SELECT 1;'}));
     await db.close();
   });
 
@@ -38,6 +36,33 @@ describe(__filename, function () {
     db.settings.queryTimeout = 100;
     await assert.doesNotReject(db._query({sql: 'DO SLEEP(0.090);'}));
     await assert.rejects(db._query({sql: 'DO SLEEP(0.110);'}));
+    await db.close();
+  });
+
+  it('queries run concurrently and are queued when pool is busy', async function () {
+    const connectionLimit = 10;
+    const db = new mysql.Database({...databases.mysql, connectionLimit});
+    await db.init();
+    // Set the query duration high enough to avoid flakiness on slow machines but low enough to keep
+    // the overall test duration short.
+    const queryDuration = 100;
+    db.settings.queryTimeout = queryDuration + 100;
+    const enqueueQuery = () => db._query({sql: `DO SLEEP(${queryDuration / 1000});`});
+
+    // Reduce test flakiness by using slow queries to warm up the pool's connections.
+    await Promise.all([...Array(connectionLimit)].map(enqueueQuery));
+
+    // Time how long it takes to run just under 2 * connectionLimit queries.
+    const nQueries = 2 * connectionLimit - 1;
+    const start = Date.now();
+    await Promise.all([...Array(nQueries)].map(enqueueQuery));
+    const duration = Date.now() - start;
+
+    const wantDurationLower = Math.ceil(nQueries / connectionLimit) * queryDuration;
+    assert(duration >= wantDurationLower, `took ${duration}ms, want >= ${wantDurationLower}ms`);
+    const wantDurationUpper = wantDurationLower + queryDuration;
+    assert(duration < wantDurationUpper, `took ${duration}ms, want < ${wantDurationUpper}ms`);
+
     await db.close();
   });
 });